refactor: extract HTTP types/enums/constants to companion module (CDMCH-55)
